changed operator and movements class to operator_file.py
